% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newAuth.R
\name{check_auth}
\alias{check_auth}
\title{check_auth}
\usage{
check_auth(auth)
}
\arguments{
\item{auth}{type string to be validated with the twitter api}
}
\value{
returns boolean, False if the bearer token is not real. True if the bearer token is real
}
\description{
helper function, validates that key passed into newAuth with the Twitter API
}
