% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tweets.R
\name{clean_tweets}
\alias{clean_tweets}
\title{clean_tweets}
\usage{
clean_tweets(tweets, options = "all")
}
\arguments{
\item{tweets}{a character vector of tweets from a data frame or normal vector}

\item{options}{either a numeric vector or string "all"
- 'all' runs all of the cleaning functions
- numeric vector with range of 1 to 6. 1 remove links. 2 removes new line tabs.
3 removes numbers. 4 removes punctuation. 5 removes stop words.
6 puts the string to lower case.}
}
\value{
a data frame of cleaned tweets
}
\description{
Provides the users to customize how to clean the twitter data
}
