% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newAuth.R
\name{get_key}
\alias{get_key}
\title{get_key}
\usage{
get_key(key)
}
\arguments{
\item{key}{object of class auth_key}
}
\value{
function returns the bearer token of the auth_key
}
\description{
returns the bearer key of an object of auth_key, pays homage to the get functions in other OOP langauges
}
