% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_list_files.R
\name{get_list_files}
\alias{get_list_files}
\title{get_list_files}
\usage{
get_list_files(path)
}
\arguments{
\item{path}{a string to a local directory to read}
}
\value{
a complied data frame from the read files
}
\description{
reads through a path and reads all of the files and creates
a complied data frame from the files
}
