% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tweet_history_ids.R
\name{get_tweet_history_ids}
\alias{get_tweet_history_ids}
\title{get_tweet_history_ids}
\usage{
get_tweet_history_ids(handles, auth_key)
}
\arguments{
\item{handles}{a character vector of all accounts a user wants to collect}

\item{auth_key}{an object of the auth_key class, will be used as the bearer token}
}
\value{
the complete data frame of all account's timeline
}
\description{
get the history of account handles in a vector
}
