% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tweets.R
\name{make_source}
\alias{make_source}
\title{make.source}
\usage{
make_source(input)
}
\arguments{
\item{input}{either vector or VCorpus}
}
\value{
a vector casted to a Volatile Corpus for additional cleaning

input is returned if input is already a dataframe
}
\description{
checks to see if the argument is of class VSource
}
