% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newAuth.R
\name{newAuth}
\alias{newAuth}
\title{new_auth}
\usage{
newAuth(auth_key)
}
\arguments{
\item{auth_key}{argument of type character, supposed to be bearer token compatible with the twitter api}
}
\value{
object of class authKey
}
\description{
creates an S4 object of type authKey
}
