% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tweet_history_ids.R
\name{start_pagination}
\alias{start_pagination}
\title{start_pagination}
\usage{
start_pagination(headers, id)
}
\arguments{
\item{headers}{contains the bearer token to access the twitter api}

\item{id}{the id of the twitter account an individual wants to collect from}
}
\value{
a combined data frame of the user's twitter history
}
\description{
helper function, collects all of the tweets from the twitter api that are stored in separate pages
}
